Adding warnings and fixes, CLI11 target, support for VIM Syntastic

+++ tests/HelpersTest.cpp
@@ -7,7 +7,7 @@
 
 TEST(Split, SimpleByToken) {
     auto out = CLI::detail::split("one.two.three", '.');
-    ASSERT_EQ(3, out.size());
+    ASSERT_EQ((size_t) 3, out.size());
     EXPECT_EQ("one", out.at(0));
     EXPECT_EQ("two", out.at(1));
     EXPECT_EQ("three", out.at(2));
@@ -15,13 +15,13 @@ TEST(Split, SimpleByToken) {
 
 TEST(Split, Single) {
     auto out = CLI::detail::split("one", '.');
-    ASSERT_EQ(1, out.size());
+    ASSERT_EQ((size_t) 1, out.size());
     EXPECT_EQ("one", out.at(0));
 }
 
 TEST(Split, Empty) {
     auto out = CLI::detail::split("", '.');
-    ASSERT_EQ(1, out.size());
+    ASSERT_EQ((size_t) 1, out.size());
     EXPECT_EQ("", out.at(0));
 }
